chime: KernelSU: Bundle in the official Manager for KernelSU
Source: https://github.com/tiann/KernelSU/releases/tag/v0.7.0
MD5: 6e358d9b02f14dde65394efe15446d71

Given the nature of the KernelSU driver module it is ABSOLUTELY IMPERATIVE that there is a manager bundled in that can protect the user from themselves. This is only here as additional security and more importantly as insurance.
The user can choose to get rid of it, or they can leave it as is.

Signed-off-by: itsHanibee <itshanibee@gmail.com>